move sidebar and Dashboard header to util so other components can reuse
